[Fix] message 독립
- 개별 message list 컴포넌트로 분리
- MessageValues 타입 선언 위치 이동
- link url ROUTE 상수화
