+++ front/pages/messages/index.tsx
@@ -1,20 +1,13 @@
 import type { NextPage } from "next";
 import { GetServerSideProps } from "next";
-import Link from "next/link";
 import { useState } from "react";
 import _ from "lodash";
 import { useQuery, useMutation, useQueryClient } from "react-query";
 import * as Api from "@api";
 import { CommonResponseDto as CommonResponse } from "@modelTypes/commonResponseDto";
 import { messageKeys } from "@utils/queryKey";
-import {
-  MAIN_COLOR,
-  ACCENT_COLOR,
-  GRAY_COLOR,
-  GREEN_COLOR,
-  ROUTE,
-} from "@utils/constant";
-import { AiOutlineCheck } from "react-icons/ai";
+import { MAIN_COLOR, ACCENT_COLOR, ROUTE } from "@utils/constant";
+
 import {
   ContentContainer,
   TitleContainer,
@@ -25,18 +18,12 @@ import {
   Label,
   DeleteBtn,
   Body,
-  List,
-  MessageContainer,
-  Message,
-  SettingBtn,
-  CheckBtn,
-  CheckBtnText,
-  Time,
   MoreBtn,
 } from "@messagesComp/MessagesPage.style";
 import Container from "@container/Container";
 import Modal from "@modal/Modal";
-import CheckModal, { MessageValues } from "@messagesComp/CheckModal";
+import CheckModal from "@messagesComp/CheckModal";
+import Messages, { MessageValues } from "@messagesComp/Messages";
 
 interface MessageResponse extends CommonResponse {
   alarms: MessageValues[];
@@ -168,34 +155,13 @@ const MessageListPage: NextPage = () => {
             </Header>
             <Body $scrollColor={ACCENT_COLOR}>
               {messages.map((message) => (
-                <List key={message.id}>
-                  <input
-                    type="checkbox"
-                    id={message.id}
-                    name="messages"
-                    checked={selectedMessagesId.includes(message.id)}
-                    onChange={selectMessageHandler}
-                  />
-                  <MessageContainer $borderColor={MAIN_COLOR}>
-                    <Message>
-                      약을 먹을 시간입니다!
-                      <br />
-                      {message.pill_name}
-                    </Message>
-                    <Link href={`/messages/setting/${message.pillBookmarkId}`}>
-                      <SettingBtn $btnColor={ACCENT_COLOR}>
-                        설정하러 가기
-                      </SettingBtn>
-                    </Link>
-                    <CheckBtn
-                      disabled={message.check}
-                      onClick={() => checkBtnClickHandler(message)}
-                      $btnColor={message.check ? GREEN_COLOR : GRAY_COLOR}>
-                      <CheckBtnText>복약 여부</CheckBtnText> <AiOutlineCheck />
-                    </CheckBtn>
-                    <Time $txtColor={GRAY_COLOR}>{message.time}</Time>
-                  </MessageContainer>
-                </List>
+                <Messages
+                  key={message.id}
+                  message={message}
+                  selectedMessagesId={selectedMessagesId}
+                  selectMessageHandler={selectMessageHandler}
+                  checkBtnClickHandler={checkBtnClickHandler}
+                />
               ))}
             </Body>
             <MoreBtn
